Updated header size for post title and added git instructions
The site is now using h2 for post titles, but the default size for
these was too big.  A little CSS and presto, right size.  Still aligns
with the navigation too.

Also, added text explaining I manage the site with git, with links to
the instructions I used and the solution to the problem I encountered.