New title image and updates to first post
I updated the title graphic with the word 'good.' so that the error
message was centered.  I'm including the GIMP image used for future
reference, where before the image came from a wallpaper website.

Also, I've updated the text with it's close to final copy.  Added
links to useful design resources.  Also included new text in the
footer.