Reinstated scrolling code blocks
Refactoring to using Pure caused the scrolling code blocks to start
wrapping due to the <pre> blocks CSS being changed.

After playing with many CSS attributes for the pre blocks and other
divs, I never could get it to work.  Instead I've created a custom
CSS class "codeblock" that has the desired attributes.

All <pre> are now replaced with <div class="codeblock">.  It's not
ideal, but it works for now.