Fix codeblock widths.
I thought I had this fixed, but just noticed that posts with code blocks were out of whack, stretching
to fit the content rather than letting the overflow work.

I was able to fix this width in all browsers except Epic by setting the width to 100% and the
overflow to scroll.  However, Epic required that the set_codeblock_widths be used as well.

The only way to get the proper width was to record the width of the #post element prior to
any text being loaded.

Tested in IE10, Chrome, Epic, and Safari.