C++ Coding Standard article and minor fixes
Added a new article!  My personal C++ coding standards cultivated from various places around
the Interwebs.  The article is a standalone page and I have a small introductory paragraph on
the main blog that links to it.  I rather like the design of the coding standard doc.

Also fixed a spelling error in the Last Configuration article.

Included are the Multimarkdown docs used to author the docs.  I’m using a new
editor, MultiMarkdown Composer 2, to write the articles.  I like it better than Mou.