Redoing the flex theme
I had to abandon the flexbox based theme I was going for because of an anomoly with long lines.  I spent a couple of days looking at it and trying things, and never could get it to work properly.

Instead, I am using the flex theme from http://blog.alexandrevicenzi.com/, an open source theme for Pelican.  I'm stripping it down to just html for now and putting in sample data, but so far it looks great, and it handles the long lines from code blocks really well.

At some point in the future, I will convert it to an AMP page for faster loading.

I'm including the latest version of the flexbox stuff just before I abandoned it, for future reference if desired.